New Company Insertion from CSV parser with Full Flexibility
NOTE: I have mailed, sample CSV file.... To see the reflection of
parsing on home page, you need to make one small change.... Replace
"AutoSoft Corp." in function getHtmlForHierarchy() (in HomeController)
with the 'companyName' that you will provide during submission...
Contact me for more details..

Following are the details for the changes...

Three new functions are added in HomeController...
1. processAdminPanelNewCompany() is the main function which will first
process the file into 2D array and will call the recursive function to
parse the file
2. addNewCompanyNode() is the recursive function for parsing CSV
3.findChildNodeOfCurrentNode() is an helper function to obtain the
children of current node.

Changes in routes file is one single line which routes the parsing
request to function in HomeController.

changes in app.php is for new Service package added.

Changes in maincontent.blad.php is for html code for adminPanel
block....